tt-02-django-stripe-docker/requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
django-environ
djangorestframework
stripe==3.5.0
#django_heroku
#gunicorn